+++ wo/core/variables.py
@@ -145,12 +145,21 @@ class WOVariables():
                          "10.3/debian {codename} main"
                          .format(codename=wo_platform_codename))
 
+    if not wo_distro == 'raspbian':
+        if (not wo_platform_codename == 'jessie'):
+            wo_mysql = ["mariadb-server", "percona-toolkit",
+                        "python3-mysqldb", "mariadb-backup"]
+        else:
+            wo_mysql = ["mariadb-server", "percona-toolkit",
+                        "python3-mysql.connector"]
+    else:
+        wo_mysql = ["mariadb-server", "percona-toolkit",
+                    "python3-mysqldb"]
+
     if wo_platform_codename == 'jessie':
         wo_mysql_client = ["mariadb-client", "python3-mysqldb"]
     else:
         wo_mysql_client = ["mariadb-client", "python3-mysql.connector"]
-
-
 
     wo_fail2ban = ["fail2ban"]
     wo_clamav = ["clamav", "clamav-freshclam"]